Moved existing Item Tables implementation into a v1 sub-directory.
Added a v2 sub-directory with stub files.
