+++ src/components/item-list/v1/ItemTable.svelte
