+++ src/components/item-list/v1/ItemTableCell.svelte
