+++ src/sheets/character/parts/CharacterConditions.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import ItemTable from 'src/components/item-list/ItemTable.svelte';
-  import ItemTableColumn from 'src/components/item-list/ItemTableColumn.svelte';
-  import ItemTableHeaderRow from 'src/components/item-list/ItemTableHeaderRow.svelte';
+  import ItemTable from 'src/components/item-list/v1/ItemTable.svelte';
+  import ItemTableColumn from 'src/components/item-list/v1/ItemTableColumn.svelte';
+  import ItemTableHeaderRow from 'src/components/item-list/v1/ItemTableHeaderRow.svelte';
   import ConditionToggle from 'src/components/toggle/ConditionToggle.svelte';
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import type { CharacterSheetContext } from 'src/types/types';
