+++ src/sheets/character/tabs/CharacterFeaturesTab.svelte
@@ -5,12 +5,12 @@
   import ItemEditControl from '../../../components/item-list/controls/ItemEditControl.svelte';
   import ItemDuplicateControl from '../../../components/item-list/controls/ItemDuplicateControl.svelte';
   import ItemDeleteControl from '../../../components/item-list/controls/ItemDeleteControl.svelte';
-  import ItemTable from '../../../components/item-list/ItemTable.svelte';
-  import ItemTableHeaderRow from '../../../components/item-list/ItemTableHeaderRow.svelte';
-  import ItemTableRow from '../../../components/item-list/ItemTableRow.svelte';
+  import ItemTable from '../../../components/item-list/v1/ItemTable.svelte';
+  import ItemTableHeaderRow from '../../../components/item-list/v1/ItemTableHeaderRow.svelte';
+  import ItemTableRow from '../../../components/item-list/v1/ItemTableRow.svelte';
   import ItemTableFooter from '../../../components/item-list/ItemTableFooter.svelte';
-  import ItemTableCell from '../../../components/item-list/ItemTableCell.svelte';
-  import ItemTableColumn from '../../../components/item-list/ItemTableColumn.svelte';
+  import ItemTableCell from '../../../components/item-list/v1/ItemTableCell.svelte';
+  import ItemTableColumn from '../../../components/item-list/v1/ItemTableColumn.svelte';
   import ItemUseButton from '../../../components/item-list/ItemUseButton.svelte';
   import { CONSTANTS } from 'src/constants';
   import ItemName from '../../../components/item-list/ItemName.svelte';
